Time: 463 ms (23.33%), Space: 381.1 MB (40.91%) - LeetHub

+++ 3678-design-task-manager/3678-design-task-manager.cpp
@@ -1,7 +1,7 @@
 class TaskManager {
 public:
-    set<pair<long long,int>> set ;
-    map<int , pair<long long,int>> map ;
+    set<pair<int,int>> set ;
+    map<int , pair<int,int>> map ;
     TaskManager(vector<vector<int>>& tasks) {
         for(auto i : tasks) {
             int task = i[1] ;
